+++ src/callbacksCustom/postPlayerInitFirst.ts
@@ -1,8 +1,3 @@
-// PostPlayerInit will fire at the beginning of a run and upon continuing a saved run.
-
-// Note that checking for `isChildPlayer` does not work in this callback; use the
-// `POST_PLAYER_INIT_LATE` callback for that.
-
 import { getPlayerIndex, log, ModCallbackCustom } from "isaacscript-common";
 import * as nLuck from "../features/items/nLuck";
 import { mod } from "../mod";