+++ src/classes/features/mandatory/RestartOnNextFrame.ts
@@ -1,6 +1,3 @@
-// This feature is used because the game prevents you from executing a "restart" console command
-// while in the `POST_GAME_STARTED` callback.
-
 import {
   Challenge,
   ModCallback,
@@ -26,6 +23,10 @@ const v = {
   },
 };
 
+/**
+ * This feature is used because the game prevents you from executing a "restart" console command
+ * while in the `POST_GAME_STARTED` callback.
+ */
 export class RestartOnNextFrame extends MandatoryModFeature {
   v = v;