+++ src/classes/features/optional/enemies/FastGhosts.ts
@@ -1,0 +1,32 @@
+import {
+  EntityType,
+  ModCallback,
+  NpcState,
+} from "isaac-typescript-definitions";
+import { Callback } from "isaacscript-common";
+import { Config } from "../../../Config";
+import { ConfigurableModFeature } from "../../../ConfigurableModFeature";
+
+export class FastGhosts extends ConfigurableModFeature {
+  configKey: keyof Config = "FastGhosts";
+
+  // 0, 219
+  @Callback(ModCallback.POST_NPC_UPDATE, EntityType.WIZOOB)
+  postNPCUpdateWizoob(npc: EntityNPC): void {
+    this.checkSpeedUpGhost(npc);
+  }
+
+  // 0, 285
+  @Callback(ModCallback.POST_NPC_UPDATE, EntityType.RED_GHOST)
+  postNPCUpdateRedGhost(npc: EntityNPC): void {
+    this.checkSpeedUpGhost(npc);
+  }
+
+  checkSpeedUpGhost(npc: EntityNPC): void {
+    // Speed up the attack pattern of Wizoobs & Red Ghosts.
+    if (npc.State === NpcState.IDLE) {
+      // This is when they are disappeared and doing nothing.
+      npc.StateFrame = 0; // `StateFrame` decrements down from 60 to 0, so just jump ahead.
+    }
+  }
+}